+++ www/go/core/views/extjs3/Notifier.js
@@ -4,6 +4,9 @@ go.Notifier = {
 	showStatusBar: false,
 	notificationArea: null,
 	init: function(notificationArea) {
+
+		var me = this;
+
 		this.notificationArea = notificationArea;
 
 		this.addStatusIcon('upload', 'ic-file-upload');
@@ -12,8 +15,10 @@ go.Notifier = {
 			this.statusBar.add(this._icons[key]);
 		}
 		this.statusBar.doLayout();
-		this.statusBar.el.on('click', function(){
-			notificationArea.toggleCollapse();
+		this.statusBar.el.on('click', function() {
+			setTimeout(function() {
+				me.showNotifications();
+			});
 		}, this);
 
 		this.notifications = new Ext.Container({cls: 'notifications'});
@@ -97,10 +102,16 @@ go.Notifier = {
 		if(key) {
 			msg.itemId = key;
 		}
+
+		//makes it fly out
+		// msg.renderTo = this.messageCt;
+
 		var msgPanel = new Ext.Panel(msg);
+
 
 		this.notifications.add(msgPanel);
 		this.notifications.doLayout();
+
 
 
 		if(msg.removeAfter) {
@@ -112,6 +123,14 @@ go.Notifier = {
 			}, msg.removeAfter);
 		}
 		msgPanel.setPersistent = function(bool) {
+
+			if(!msgPanel.rendered) {
+				msgPanel.on("render", function() {
+					msgPanel.setPersistent(bool);
+				}, this, {single: true});
+				return msgPanel;
+			}
+
 			msgPanel.getTool('close').setVisible(!bool);
 			return msgPanel;
 		};
@@ -122,8 +141,18 @@ go.Notifier = {
 			}
 			this._messages[msg.itemId] = msgPanel;
 		}
+
+		//this.showNotifications();
 		
 		return msgPanel;
+	},
+
+	showNotifications : function() {
+		this.notificationArea.ownerCt.getLayout()['east'].slideOut();
+	},
+
+	hideNotifications : function() {
+		this.notificationArea.ownerCt.getLayout()['east'].slideIn();
 	},
 	/**
 	 * For (less obstructive) popup messages from the bottom